move components into folder
