+++ src/components/Analysis/Activity/CustomNode.vue
